+++ lib/braintree/payment_instrument_type.rb
@@ -3,8 +3,7 @@ module Braintree
     PayPalAccount = 'paypal_account'
     CreditCard = 'credit_card'
     ApplePayCard = 'apple_pay_card'
-    # NEXT_MAJOR_VERSION rename Android Pay to Google Pay
-    AndroidPayCard = 'android_pay_card'
+    GooglePayCard = 'android_pay_card'
     VenmoAccount = 'venmo_account'
     UsBankAccount = 'us_bank_account'
     VisaCheckoutCard = 'visa_checkout_card'