+++ lib/braintree/util.rb
@@ -126,6 +126,17 @@ module Braintree
       !invalid_keys.any?
     end
 
+    def self.replace_key(hash, target_key, replacement_key)
+      hash.inject({}) do |new_hash, (key, value)|
+        if value.is_a?(Hash)
+          value = replace_key(value, target_key, replacement_key)
+        end
+
+        key = replacement_key if key == target_key
+        new_hash.merge(key => value)
+      end
+    end
+
     def self._flatten_valid_keys(valid_keys, namespace = nil)
       valid_keys.inject([]) do |result, key|
         if key.is_a?(Hash)